requirements.txt
flask
psutil
speedtest-cli
fabric<2.0